refactor: 온도 입력 관련 폴더 구조 변경
# 선행된 리팩토링
1. Modal: 모달은 특정한 화면을 화면 위에 띄우는, 사용자의 이목을 집중시키는 그래픽 인터페이스입니다.
  - 문제 상황
    1. 기존 모달은 재사용 가능한 모달이 아닌 Confirm 컴포넌트를 화면에 표시하기 위한 컴포넌트로 동작.
    2. 오직 화면에 표시하는 기능을 수행해야하지만 전달 받는 error, submit 등 특정 컴포넌트만을 위한 컴포넌트로 동작하게 됨.
  - 개선사항
    1. Footer에서 새로 구현해 사용 중인 Modal을 이용. 해당 Modal은 오로지 전달 받은 children 요소를 body의 modal-root 요소로 렌더링함.

# 폴더 구조 변경

- 페이지 단위 (pages)
  1. InputTempMachines, InputTempFoods: 제품 온도 입력 페이지는 페이지 단위로 pages 폴더로 이동합니다. 공통된 역할인 input-temp 폴더로 묶고, food, machine 폴더에서 관리합니다.
  2. select-foods, select-machines: 마찬가지로 select-products 폴더로 묶은 뒤 food, machine 폴더로 관리합니다.

- 위젯 단위 (widgets)
  1. InputTempForm: 온도를 입력하는 기능을 수행하는 UI 요소로 특정 페이지(기기 및 식품 입력)에서만 사용되는 재사용 컴포넌트로 widgets 폴더로 이동합니다.

- 기능 단위 (features)
  1. InputTemp: react-hook-from을 사용한 defaultValues를 기반으로 온도를 입력하는 기능을 제공하므로 features 폴더로 이동합니다.
  2. ManagerList: 매니저를 선택하는 기능으로 컴포넌트 이름을 SelectManager로 변경 후 features 폴더로 이동합니다. CreatableSelect 라이브러리에 제공하는 customStyles는 model에서 관리합니다.
  3. alert(error-alert, success-alert): 기존 Modal => Confirm으로 표시하던 성공 알림과 재시도 알림을 각 상황에 맞는 컴포넌트를 통해 사용자에게 표시하는 기능으로 변경했습니다.

- 엔티티 단위 (entities)
  1. useCustomMachines, Foods, useCustomProducts: 기존 온도 입력 페이지에서 사용되던 데이터 쿼리를 받아오는 함수와제출 함수를 entities 폴더로 이동합니다. 쿼리 관련 함수는 query, 제출 로직에 관련된 함수는 action에서 관리합니다.
  2. manager state: 온도 입력에 사용되는 manager state를 관리하는 스토어를 entities 폴더로 이동합니다.
