+++ src/entities/custom-products/index.js
@@ -1,0 +1,3 @@
+export { useCustomMachines } from "./model/query";
+export { useCustomFoods } from "./model/query";
+export { useCustomProducts } from "./model/action";
