+++ src/entities/manager/index.js
@@ -1,0 +1,2 @@
+export { useManagerState } from "./model/store";
+export { useSetManager } from "./model/store";
