+++ src/entities/manager/model/store.js
