--- src/entities/products/index.js
@@ -1,5 +1,0 @@
-export { useMachines } from "./query";
-export { useFoods } from "./query";
-export { useCustomProducts } from "./store";
-export { useCustomProductsActions } from "./store";
-export { useResetCustomProducts } from "./store";
