+++ src/entities/select-products/model/store.js
