+++ src/features/alert/index.js
@@ -1,0 +1,2 @@
+export { default as SuccessAlert } from "./success-alert/SuccessAlert";
+export { default as ErrorAlert } from "./error-alert/ErrorAlert";
