+++ src/features/alert/styles.module.css
@@ -1,0 +1,32 @@
+.section {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  font-family: "Pretendard";
+  z-index: 999;
+  width: 400px;
+  height: 200px;
+  color: var(--navbar-background-color);
+  background-color: var(--screen-background-color);
+  border: 2px solid var(--navbar-background-color);
+  border-radius: 8px;
+  overflow: hidden;
+}
+
+.content {
+  padding: 15px;
+}
+
+.header {
+  font-size: 1.3rem;
+  font-weight: 700;
+}
+
+.description {
+  flex: 1;
+  margin-top: 16px;
+}
+
+.container {
+  display: flex;
+}
