+++ src/features/input-temp/index.js
@@ -1,0 +1,1 @@
+export { default as InputTemp } from "./ui/InputTemp";
