+++ src/features/select-manager/index.js
@@ -1,0 +1,1 @@
+export { default as SelectManager } from "./ui/SelectManager";
