+++ src/features/select-manager/model/style.js
