+++ src/pages/input-temp/index.js
@@ -1,0 +1,2 @@
+export { default as InputMachineTemp } from "./machine/InputMachineTemp";
+export { default as InputFoodTemp } from "./food/InputFoodTemp";
