+++ src/pages/input-temp/styles.module.css
@@ -1,51 +1,45 @@
 .section {
   display: flex;
   flex-direction: column;
+  justify-content: space-between;
   height: 100%;
 }
 
-.wrapper {
-  flex: 1;
-  display: flex;
-  overflow: auto;
-  flex-direction: column;
-  width: 100%;
-}
-
 .title {
-  position: sticky;
   flex: 0;
+  position: sticky;
+  z-index: 990;
   width: 100%;
-  z-index: 999;
   top: 0;
   text-align: center;
   padding: 0.7rem 0;
   font-size: xx-large;
-  font-weight: 700;
+  font-weight: bold;
   margin-bottom: 0.5rem;
   background-color: var(--navbar-background-color);
   color: var(--text-basic-color);
 }
 
-.form {
+.empty {
   display: flex;
-  flex-direction: column;
-  padding: 2rem;
-  flex: 1;
+  justify-content: center;
+  margin-top: 12rem;
+  color: var(--navbar-background-color);
+  font-size: x-large;
 }
 
-.products {
+.form {
   display: flex;
+  flex: 1;
   flex-direction: column;
-  gap: 1rem;
+  overflow: auto;
+  padding: 2rem 2.5rem;
 }
 
 .buttons {
   display: flex;
-  position: sticky;
-  bottom: 0;
-  transform: translateY(2%);
   justify-content: space-around;
+  transform: translateY(2%);
   margin-top: 1rem;
   font-weight: bold;
 }
