--- src/pages/select-foods/index.js
@@ -1,1 +1,0 @@
-export { default as CustomFoods } from "./ui/CustomFoods";
