--- src/pages/select-machines/index.js
@@ -1,1 +1,0 @@
-export { default as CustomMachines } from "./ui/CustomMachines";
