+++ src/pages/select-products/index.js
@@ -1,0 +1,2 @@
+export { default as SelectMachines } from "./machine/SelectMachines";
+export { default as SelectFoods } from "./food/SelectFoods";
