+++ src/pages/select-products/machine/SelectMachines.jsx
@@ -1,22 +1,25 @@
 import React, { useEffect } from "react";
 import styles from "../styles.module.css";
 import { CustomProduct } from "../../../widgets/custom-product";
-import { useFoods, useResetCustomProducts } from "../../../entities/products";
-import { useSubmitActions, useSubmitStates } from "../../../entities/ui-state";
+import {
+  useMachines,
+  useResetCustomProducts,
+} from "../../../entities/select-products";
+import { useSubmitStates, useSubmitActions } from "../../../entities/ui-state";
 import { Button } from "../../../shared/ui/Button";
 import { Banner } from "../../../shared/ui/Banner";
 import { BarSpinner } from "../../../shared/ui/LoadingSpinner";
 import { useNavigateToMain } from "../../../shared/lib/hooks";
 
-export default function CustomFoods() {
+export default function SelectMachines() {
   const { handleClick } = useNavigateToMain();
   const { success } = useSubmitStates();
   const { resetState } = useSubmitActions();
 
   const {
-    productsQuery: { isLoading, error, data: foods },
+    productsQuery: { isLoading, error, data: machines },
     handleSubmitProducts,
-  } = useFoods();
+  } = useMachines();
 
   const resetProducts = useResetCustomProducts();
 
@@ -29,31 +32,31 @@ export default function CustomFoods() {
 
   return (
     <section className={styles.section}>
-      <div className={styles.title}>식품 선택</div>
+      <div className={styles.title}>기기 선택</div>
       <div className={styles.wrapper}>
         <form
-          id="customFoods"
+          id="customMachine"
           className={styles.form}
           onSubmit={handleSubmitProducts}
         >
-          {success && <Banner text={"식품 선택이 완료되었습니다."} />}
+          {success && <Banner text={"기기 선택이 완료되었습니다."} />}
           <div className={styles.products}>
             {error && <p>{error}</p>}
             {isLoading && <BarSpinner />}
-            {foods &&
-              foods.map((food) => (
-                <div key={food.id}>
+            {machines &&
+              machines.map((machine) => (
+                <div key={machine.id}>
                   <CustomProduct
-                    value={food}
-                    checkedIt={food.isChecked === "true" ? true : false}
+                    value={machine}
+                    checkedIt={machine.isChecked === "true" ? true : false}
                   />
                 </div>
               ))}
           </div>
         </form>
         <div className={styles.buttons}>
-          <Button text={"저장"} form={"customFoods"} />
-          <Button text={"취소"} handleFunction={handleClick} />
+          <Button text={"저장"} type={"submit"} form={"customMachine"} />
+          <Button text={"취소"} onClick={handleClick} />
         </div>
       </div>
     </section>
