+++ src/shared/ui/Button/ui/Button.jsx
@@ -1,13 +1,22 @@
 import React from "react";
 import styles from "../styles.module.css";
 
-export default function Button({ text, handleFunction, type, form, disabled }) {
+export default function Button({
+  text,
+  onClick,
+  type,
+  form,
+  disabled,
+  size = "middle",
+}) {
   return (
     <button
-      className={styles.button}
+      className={`${styles.buttonCommon} ${
+        size === "middle" ? styles.button : styles.smallButton
+      }`}
       type={type && type}
       form={form && form}
-      onClick={handleFunction && handleFunction}
+      onClick={onClick && onClick}
       disabled={disabled ? true : false}
     >
       {text}
