+++ src/widgets/input-temp-form/index.js
@@ -1,0 +1,1 @@
+export { default as InputTempForm } from "./ui/InputTempForm";
