+++ src/widgets/input-temp-form/styles.module.css
@@ -29,3 +29,22 @@
   font-size: 0.8rem;
   color: #eed202;
 }
+
+.empty {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  text-align: center;
+  font-size: 1.3rem;
+  color: var(--navbar-background-color);
+}
+
+.navigate {
+  font-size: 1.1rem;
+  font-weight: 700;
+  background-color: var(--main-background-color);
+  width: fit-content;
+  margin: 0 auto;
+  padding: 0.7rem 2rem;
+  border-radius: 8px;
+}
