+++ src/widgets/input-temp-form/ui/InputTempForm.jsx
@@ -1,8 +1,9 @@
+import styles from "../styles.module.css";
+import { Link } from "react-router-dom";
 import { useForm } from "react-hook-form";
-import InputTemp from "./InputTemp";
-import styles from "./InputTempForm.module.css";
+import { InputTemp } from "../../../features/input-temp";
 
-export default function InputTempForm({ onSubmit, products }) {
+export default function InputTempForm({ onSubmit, products, pageLocation }) {
   const {
     register,
     handleSubmit,
@@ -20,6 +21,17 @@ export default function InputTempForm({ onSubmit, products }) {
       })),
     },
   });
+
+  if (products.length === 0) {
+    return (
+      <section className={styles.empty}>
+        <p>먼저 사용할 제품을 선택해주세요.</p>
+        <Link className={styles.navigate} to={`/select/${pageLocation}`}>
+          선택하러 가기
+        </Link>
+      </section>
+    );
+  }
 
   return (
     <form
